requirements.txt
langchain==0.1.12
langchain-community==0.0.27
langchain-core==0.1.32
langchain-openai==0.0.8
openai==1.14.0
python-dotenv==1.0.1
faiss-cpu==1.7.4
streamlit==1.45.1
beautifulsoup4==4.12.3
lxml==5.2.0
pypdf==4.1.0
unstructured==0.12.4
openpyxl==3.1.2
pandas==2.2.3

# Testing dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0